Fixed block synchronization stalling issue
- Wrong binding of events caused the event handler to exit after the first received event.

- When `BlockOk` or `TransactionOk` messages indicate rejection, close the stream instead of a reset.

- Removed peer as a broadcaster if it's no longer connected to the local peer

- Fixed broken tests.
