+++ blockchain/process.go
@@ -506,7 +506,10 @@ commit:
 		b.eventEmitter.Emit(core.EventNewBlock, block, chain.ChainReader())
 	}
 
-	b.log.Info("Block has been processed", "BlockNo", block.GetNumber())
+	b.log.Info("Block has been processed",
+		"BlockNo", block.GetNumber(),
+		"BlockHash", block.GetHash().SS(),
+		"ChainID", chain.GetID().SS())
 
 	return chain, nil
 }
